refactor: separated utility functions from components
