+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export { textFormatter } from './formatter';
+export { getFilteredProducts } from './product-filters';
